refactor

+++ code/Interpreter.hpp
@@ -13,15 +13,15 @@ public:
 private:
 	void PrintHelp();
 
-	const size_t CHILDREN_LIMIT = 15000000000;
+	const long CHILDREN_LIMIT = 15000000000;
 
 	int argumentsNo;
 	char **arguments;
 
 	bool manualTests;
-	size_t testsNo;
-	size_t pupilsNo;
-	size_t maxMark;
+	long testsNo;
+	long pupilsNo;
+	long maxMark;
 };
 
 #endif
